feat: hook useLoadAquariums
