+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useAquariumDetails } from './useAquariumDetails';
 export { useCreateAquarium } from './useCreateAquarium';
+export { useLoadAquariums } from './useLoadAquariums';
 export { useResetParams } from './useResetParams';
